pyproject.toml
[project]
name = "espresso-service"
version = "0.1.0"
description = "ESPResso carbon footprint prediction service for Avelero DPP platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.28.0",
    "lightgbm>=4.5.0",
    "joblib>=1.4.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "pydantic-settings>=2.7.0",
    "cachetools>=5.5.0",
    "structlog>=24.4.0",
    "scikit-learn>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]
db = [
    "psycopg2-binary>=2.9.0",
]

[tool.setuptools.packages.find]
include = ["app*", "espresso_models*"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
